Ensure QR Code not shown in react-native-web apps when on mobile browsers
Previously the modal would render a QR code anytime react-native was running on web even if it
was running on mobile Safari / Chrome etc. This fixes it by checking if the user agent device type
corresponds to a mobile or tablet device (iOS or Android).

Test Plan: Open a react-native-web app with WalletConnect on mobile Safari and ensure a grid
of wallet options is shown instead of a QR Code.